backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
bson==0.5.10
httpx==0.25.2
cryptography